# specs/000-program/contracts/openapi.yaml
openapi: 3.0.3
info:
  title: 肇庆市"好房子"在线选房与户型智能评估系统 API
  version: 1.0.0
  description: |
    全局契约（先于实现，宪法第二条 5）。所有响应包装为 `ApiResponse`：
    `{code, message, traceId, data}`；`code=0` 成功，错误码见 specs/000 + plan.md §3.1。
    鉴权：`Authorization: Bearer <JWT>`；`/api/admin/**` 需 ADMIN/CONSULTANT，写接口均需鉴权（NFR-06）。
    分页：`page`（1 起）、`size`（≤100，默认 20）；排序 `sort=totalPrice,asc`。
servers:
  - url: http://localhost:8080/api
    description: 本地
...
paths:
...
  /house-types:
    get:
      tags: [house-type]
      summary: 户型分页与筛选（FR-11）
      parameters:
        - { $ref: "#/components/parameters/page" }
        - { name: projectId, in: query, schema: { type: integer, format: int64 } }
        - { name: rooms, in: query, schema: { type: integer } }
        - { name: minArea, in: query, schema: { type: number } }
        - { name: maxArea, in: query, schema: { type: number } }
        - { name: orientation, in: query, schema: { type: string, enum: [S, SE, SW, N, E, W, NS] } }
        - { name: maxTotalPrice, in: query, schema: { type: number } }
        - { name: sort, in: query, schema: { type: string, example: "score,desc" } }
      responses:
        "200":
          description: "Page<HouseTypeCardVO>，含 latestScore/level 徽标字段"
          content: { application/json: { schema: { $ref: "#/components/schemas/ApiResponse" } } }
  /house-types/{id}:
    get:
      tags: [house-type]
      summary: 户型详情（几何 + 参数 + 房间构件，满足 geometry.schema.json）
      parameters: [ { name: id, in: path, required: true, schema: { type: integer, format: int64 } } ]
      responses: { "200": { description: HouseTypeDetailVO }, "400": { description: "40420 户型不存在" } }
  /house-types/{id}/rooms:
    get: { tags: [house-type], summary: 房间构件列表, parameters: [ { name: id, in: path, schema: { type: integer, format: int64 } } ], responses: { "200": { description: List<RoomVO> } } }
...
  /favorites:
    get: { tags: [selection], security: [ { bearer: [] } ], summary: 我的收藏（户型/房源）, parameters: [ { name: type, in: query, schema: { type: string, enum: [HOUSE_TYPE, HOUSE] } } ], responses: { "200": { description: List<FavoriteVO> } } }
    post: { tags: [selection], security: [ { bearer: [] } ], summary: 新增收藏（幂等，上限 50）, requestBody: { content: { application/json: { schema: { type: object, required: [targetType, targetId], properties: { targetType: { type: string }, targetId: { type: integer, format: int64 } } } } } }, responses: { "200": { description: OK }, "400": { description: "40914 超过收藏上限" } } }
...
  /selection/locks:
    post:
      tags: [selection]
      security: [ { bearer: [] } ]
      summary: 模拟选房（锁定房源，FR-34/35，TTL 10min）
      requestBody: { required: true, content: { application/json: { schema: { type: object, required: [houseId, confirm], properties: { houseId: { type: integer, format: int64 }, confirm: { type: boolean, description: "必须为 true，否则 40013" } } } } } }
      responses:
        "200": { description: "data:{lockNo,intentionNo,expireAt,renewCount}" }
        "409": { description: "40910 已被他人锁定（data.alternatives ≥3 条同户型备选）/ 40911 已售或状态不可选" }
  /selection/locks/current:
    get: { tags: [selection], security: [ { bearer: [] } ], summary: 我的有效锁（AC-22 倒计时基准）, responses: { "200": { description: "data: null 或 LockVO" } } }
...
  /selection/locks/{no}/cancel:
    post: { tags: [selection], security: [ { bearer: [] } ], summary: 取消锁定并释放房源, parameters: [ { name: no, in: path, schema: { type: string } } ], responses: { "200": { description: OK } } }
  /selection/locks/{no}/convert:
    post: { tags: [selection], security: [ { bearer: [] } ], summary: 转意向（LOCKED→RESERVED，24h）, parameters: [ { name: no, in: path, schema: { type: string } } ], responses: { "200": { description: "data:{intentionNo, expireAt}" } } }
  /evaluations:
    post:
      tags: [evaluation]
      security: [ { bearer: [] } ]
      summary: 发起评估（FR-50），可匿名预览（不落库）需 query `preview=true`
      parameters: [ { name: preview, in: query, schema: { type: boolean, default: false } } ]
      requestBody: { required: true, content: { application/json: { schema: { $ref: "#/components/schemas/EvaluateCmd" } } } }
      responses: { "200": { description: "EvaluationResultVO（total/level/dimensions[]/metrics[]/suggestions[]/setVersion）" } }
    get: { tags: [evaluation], security: [ { bearer: [] } ], summary: 我的评估历史, parameters: [ { name: houseTypeId, in: query, schema: { type: integer, format: int64 } } ], responses: { "200": { description: Page<EvaluationBriefVO> } } }
  /evaluations/{id}:
    get: { tags: [evaluation], summary: 读取评测快照（不重算，AC-32）, parameters: [ { name: id, in: path, schema: { type: integer, format: int64 } } ], responses: { "200": { description: EvaluationSnapshotVO }, "400": { description: "40302 非本人记录" } } }
  /house-types/{id}/evaluation/latest:
    get: { tags: [evaluation], summary: 某户型最近一次评估（列表徽标复用）, parameters: [ { name: id, in: path, schema: { type: integer, format: int64 } } ], responses: { "200": { description: "data: EvaluationBriefVO | null" } } }
...
  /appointments:
    post: { tags: [appointment], security: [ { bearer: [] } ], summary: 提交预约看房（FR-90）, requestBody: { content: { application/json: { schema: { $ref: "#/components/schemas/AppointmentCmd" } } } }, responses: { "200": { description: "data:{id,status:PENDING}" }, "400": { description: "40018 时段已满 / 40019 与未到访预约冲突" } } }
    get: { tags: [appointment], security: [ { bearer: [] } ], summary: 我的预约, parameters: [ { name: status, in: query, schema: { type: string } } ], responses: { "200": { description: List<AppointmentVO> } } }
...
components:
  securitySchemes:
    bearer: { type: http, scheme: bearer, bearerFormat: JWT }
  parameters:
    page: { name: page, in: query, schema: { type: integer, default: 1, minimum: 1 } }
...
  schemas:
    ApiResponse:
      type: object
      required: [code, message]
      properties:
        code: { type: integer, description: "0=成功；错误码分段见 plan.md §3.1" }
        message: { type: string }
        traceId: { type: string }
        data: { nullable: true }
...
    EvaluateCmd:
      type: object
      required: [houseTypeId]
      properties:
        houseTypeId: { type: integer, format: int64 }
        houseId: { type: integer, format: int64, description: "传入则纳入楼层/遮挡/价格修正" }
        templateCode: { type: string, enum: [GENERAL, FAMILY_3, FAMILY_4, MULTI_GEN, SENIOR_COUPLE, INVESTOR], default: GENERAL }
...
    AppointmentCmd:
      type: object
      required: [projectId, visitDate, visitSlot, contactPhone, partySize]
      properties:
        projectId: { type: integer, format: int64 }
        houseTypeId: { type: integer, format: int64 }
        houseId: { type: integer, format: int64 }
        intentionNo: { type: string, description: "由选房意向带入（AC-24）" }
        visitDate: { type: string, format: date }
        visitSlot: { type: string, example: "14:00-15:00" }
        partySize: { type: integer, minimum: 1, maximum: 6 }
        contactPhone: { type: string }
        remark: { type: string, maxLength: 200 }